DeepLearningModel/requirements.txt
absl-py==0.7.0
astor==0.7.1
bleach==1.5.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
html5lib==0.9999999
javalang==0.12.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
mock==2.0.0
numpy==1.16.1
pandas==0.24.1
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
pycparser==2.19
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
tensorboard==1.7.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.1
Werkzeug==0.14.1
